+++ Tasks/AzureTestPlanV0/Common/SimpleTimer.ts
