+++ Tasks/AzureTestPlanV0/Common/ciEventLogger.ts
