+++ Tasks/AzureTestPlanV0/Common/publishAutomatedTests.ts
