+++ Tasks/AzureTestPlanV0/Invokers/maveninvoker.ts
@@ -1,7 +1,7 @@
 import { spawn } from '../testexecutor'
 import tl = require('azure-pipelines-task-lib/task');
-import utils = require('../utils');
-import constants = require('../constants');
+import utils = require('../Common/utils');
+import constants = require('../Common/constants');
 import { execMavenBuild } from '../testLibExecutor';
 
 export async function executeMavenTests(testsToBeExecuted: string[], pomFilePath?: string):Promise<number> {
